Initial pass of WooCommerce code split
