+++ includes/classes/Feature/WooCommerce/WooCommerce.php
@@ -23,7 +23,23 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class WooCommerce extends Feature {
 	/**
-	 * If enabled, receive the Orders object instance
+	 * If enabled, receive the OrdersAutosuggest object instance
+	 *
+	 * @since 4.7.0
+	 * @var null|OrdersAutosuggest
+	 */
+	public $orders_autosuggest = null;
+
+	/**
+	 * Receive the Products object instance
+	 *
+	 * @since 4.7.0
+	 * @var null|Products
+	 */
+	public $products = null;
+
+	/**
+	 * Receive the Orders object instance
 	 *
 	 * @since 4.5.0
 	 * @var null|Orders
@@ -54,172 +70,39 @@ class WooCommerce extends Feature {
 			'orders' => '0',
 		];
 
-		$this->orders = new Orders();
+		$this->orders             = new Orders( $this );
+		$this->products           = new Products( $this );
+		$this->orders_autosuggest = new OrdersAutosuggest();
 
 		parent::__construct();
 	}
 
 	/**
-	 * Index Woocommerce meta
+	 * Setup all feature filters
 	 *
-	 * @param   array $meta Existing post meta.
-	 * @param   array $post Post arguments array.
-	 * @since   2.1
-	 * @return  array
+	 * @since  2.1
 	 */
-	public function whitelist_meta_keys( $meta, $post ) {
-		return array_unique(
-			array_merge(
-				$meta,
-				array(
-					'_thumbnail_id',
-					'_product_attributes',
-					'_wpb_vc_js_status',
-					'_swatch_type',
-					'total_sales',
-					'_downloadable',
-					'_virtual',
-					'_regular_price',
-					'_sale_price',
-					'_tax_status',
-					'_tax_class',
-					'_purchase_note',
-					'_featured',
-					'_weight',
-					'_length',
-					'_width',
-					'_height',
-					'_visibility',
-					'_sku',
-					'_sale_price_dates_from',
-					'_sale_price_dates_to',
-					'_price',
-					'_sold_individually',
-					'_manage_stock',
-					'_backorders',
-					'_stock',
-					'_upsell_ids',
-					'_crosssell_ids',
-					'_stock_status',
-					'_product_version',
-					'_product_tabs',
-					'_override_tab_layout',
-					'_suggested_price',
-					'_min_price',
-					'_customer_user',
-					'_variable_billing',
-					'_wc_average_rating',
-					'_product_image_gallery',
-					'_bj_lazy_load_skip_post',
-					'_min_variation_price',
-					'_max_variation_price',
-					'_min_price_variation_id',
-					'_max_price_variation_id',
-					'_min_variation_regular_price',
-					'_max_variation_regular_price',
-					'_min_regular_price_variation_id',
-					'_max_regular_price_variation_id',
-					'_min_variation_sale_price',
-					'_max_variation_sale_price',
-					'_min_sale_price_variation_id',
-					'_max_sale_price_variation_id',
-					'_default_attributes',
-					'_swatch_type_options',
-					'_order_key',
-					'_billing_company',
-					'_billing_address_1',
-					'_billing_address_2',
-					'_billing_city',
-					'_billing_postcode',
-					'_billing_country',
-					'_billing_state',
-					'_billing_email',
-					'_billing_phone',
-					'_shipping_address_1',
-					'_shipping_address_2',
-					'_shipping_city',
-					'_shipping_postcode',
-					'_shipping_country',
-					'_shipping_state',
-					'_billing_last_name',
-					'_billing_first_name',
-					'_shipping_first_name',
-					'_shipping_last_name',
-					'_variations_skus',
-				)
-			)
-		);
-	}
-
-	/**
-	 * Make sure all loop shop post ins are IDS. We have to pass post objects here since we override
-	 * the fields=>id query for the layered filter nav query
-	 *
-	 * @param   array $posts Post object array.
-	 * @since   2.1
-	 * @return  array
-	 */
-	public function convert_post_object_to_id( $posts ) {
-		_doing_it_wrong( __METHOD__, 'This filter was removed from WooCommerce and will be removed from ElasticPress in a future release.', '4.5.0' );
-		return $posts;
-	}
-
-	/**
-	 * Index Woocommerce taxonomies
-	 *
-	 * @param   array $taxonomies Index taxonomies array.
-	 * @param   array $post Post properties array.
-	 * @since   2.1
-	 * @return  array
-	 */
-	public function whitelist_taxonomies( $taxonomies, $post ) {
-		$woo_taxonomies = [];
-
-		$product_type = get_taxonomy( 'product_type' );
-		if ( false !== $product_type ) {
-			$woo_taxonomies[] = $product_type;
+	public function setup() {
+		if ( ! function_exists( 'WC' ) ) {
+			return;
 		}
 
-		$product_visibility = get_taxonomy( 'product_visibility' );
-		if ( false !== $product_visibility ) {
-			$woo_taxonomies[] = $product_visibility;
+		$this->products->setup();
+		$this->orders->setup();
+
+		add_filter( 'ep_integrate_search_queries', [ $this, 'disallow_coupons' ], 10, 2 );
+
+		// These hooks are deprecated and will be removed in an upcoming major version of ElasticPress
+		add_filter( 'woocommerce_layered_nav_query_post_ids', [ $this, 'convert_post_object_to_id' ], 10, 4 );
+		add_filter( 'woocommerce_unfiltered_product_ids', [ $this, 'convert_post_object_to_id' ], 10, 4 );
+		add_action( 'ep_wp_query_search_cached_posts', [ $this, 'disallow_duplicated_query' ], 10, 2 );
+
+		add_action( 'pre_get_posts', [ $this, 'translate_args' ], 11, 1 );
+
+		// Orders Autosuggest feature.
+		if ( $this->is_orders_autosuggest_enabled() ) {
+			$this->orders_autosuggest->setup();
 		}
-
-		/**
-		 * Note product_shipping_class, product_cat, and product_tag are already public. Make
-		 * sure to index non-attribute taxonomies.
-		 */
-		$attribute_taxonomies = wc_get_attribute_taxonomies();
-
-		if ( ! empty( $attribute_taxonomies ) ) {
-			foreach ( $attribute_taxonomies as $tax ) {
-				$name = wc_attribute_taxonomy_name( $tax->attribute_name );
-
-				if ( ! empty( $name ) ) {
-					if ( empty( $tax->attribute_ ) ) {
-						$woo_taxonomies[] = get_taxonomy( $name );
-					}
-				}
-			}
-		}
-
-		return array_merge( $taxonomies, $woo_taxonomies );
-	}
-
-	/**
-	 * Disallow duplicated ES queries on Orders page.
-	 *
-	 * @since 2.4
-	 *
-	 * @param array    $value Original filter values.
-	 * @param WP_Query $query WP_Query
-	 *
-	 * @return array
-	 */
-	public function disallow_duplicated_query( $value, $query ) {
-		_doing_it_wrong( __METHOD__, 'This filter was removed from WooCommerce and will be removed from ElasticPress in a future release.', '4.5.0' );
-
-		return $value;
 	}
 
 	/**
@@ -591,36 +474,15 @@ class WooCommerce extends Feature {
 		return 'date';
 	}
 
-
-	/**
-	 * Returns the WooCommerce-oriented post types in admin that EP will search
-	 *
-	 * @since 4.4.0
-	 * @return mixed|void
-	 */
-	public function get_admin_searchable_post_types() {
-		$searchable_post_types = array( 'shop_order' );
-
-		/**
-		 * Filter admin searchable WooCommerce post types
-		 *
-		 * @hook ep_woocommerce_admin_searchable_post_types
-		 * @since 4.4.0
-		 * @param  {array} $post_types Post types
-		 * @return  {array} New post types
-		 */
-		return apply_filters( 'ep_woocommerce_admin_searchable_post_types', $searchable_post_types );
-	}
-
 	/**
 	 * Make search coupons don't go through ES
 	 *
 	 * @param  bool     $enabled Coupons enabled or not
 	 * @param  WP_Query $query WP Query
-	 * @since  2.1
+	 * @since  4.7.0
 	 * @return bool
 	 */
-	public function blacklist_coupons( $enabled, $query ) {
+	public function disallow_coupons( $enabled, $query ) {
 		if ( is_admin() ) {
 			return $enabled;
 		}
@@ -630,271 +492,6 @@ class WooCommerce extends Feature {
 		}
 
 		return $enabled;
-	}
-
-	/**
-	 * Allow order creations on the front end to get synced
-	 *
-	 * @since  2.1
-	 * @param  bool $override Original order perms check value
-	 * @param  int  $post_id Post ID
-	 * @return bool
-	 */
-	public function bypass_order_permissions_check( $override, $post_id ) {
-		$searchable_post_types = $this->get_admin_searchable_post_types();
-
-		if ( in_array( get_post_type( $post_id ), $searchable_post_types, true ) ) {
-			return true;
-		}
-
-		return $override;
-	}
-
-	/**
-	 * Sets woocommerce meta search fields to an empty array if we are integrating the main query with ElasticSearch
-	 *
-	 * Woocommerce calls this action as part of its own callback on parse_query. We add this filter only if the query
-	 * is integrated with ElasticSearch.
-	 * If we were to always return array() on this filter, we'd break admin searches when WooCommerce module is activated
-	 * without the Protected Content Module
-	 *
-	 * @param \WP_Query $query Current query
-	 */
-	public function maybe_hook_woocommerce_search_fields( $query ) {
-		global $pagenow, $wp, $wc_list_table, $wp_filter;
-
-		if ( ! $this->should_integrate_with_query( $query ) ) {
-			return;
-		}
-
-		/**
-		 * Determines actions to be applied, or removed, if doing a WooCommerce serarch
-		 *
-		 * @hook ep_woocommerce_hook_search_fields
-		 * @since  4.4.0
-		 */
-		do_action( 'ep_woocommerce_hook_search_fields' );
-
-		if ( 'edit.php' !== $pagenow || empty( $wp->query_vars['s'] ) || 'shop_order' !== $wp->query_vars['post_type'] || ! isset( $_GET['s'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification
-			return;
-		}
-
-		remove_action( 'parse_query', [ $wc_list_table, 'search_custom_fields' ] );
-	}
-
-	/**
-	 * Enhance WooCommerce search order by order id, email, phone number, name, etc..
-	 * What this function does:
-	 * 1. Reverse the woocommerce shop_order_search_custom_fields query
-	 * 2. If the search key is integer and it is an Order Id, just query with post__in
-	 * 3. If the search key is integer but not an order id ( might be phone number ), use ES to find it
-	 *
-	 * @param WP_Query $wp WP Query
-	 * @since  2.3
-	 */
-	public function search_order( $wp ) {
-		if ( ! $this->should_integrate_with_query( $wp ) ) {
-			return;
-		}
-
-		global $pagenow;
-
-		$searchable_post_types = $this->get_admin_searchable_post_types();
-
-		if ( 'edit.php' !== $pagenow || empty( $wp->query_vars['post_type'] ) || ! in_array( $wp->query_vars['post_type'], $searchable_post_types, true ) ||
-			( empty( $wp->query_vars['s'] ) && empty( $wp->query_vars['shop_order_search'] ) ) ) {
-			return;
-		}
-
-		// phpcs:disable WordPress.Security.NonceVerification, WordPress.Security.ValidatedSanitizedInput
-		if ( isset( $_GET['s'] ) ) {
-			$search_key_safe = str_replace( array( 'Order #', '#' ), '', wc_clean( $_GET['s'] ) );
-			unset( $wp->query_vars['post__in'] );
-			$wp->query_vars['s'] = $search_key_safe;
-		}
-		// phpcs:enable WordPress.Security.NonceVerification, WordPress.Security.ValidatedSanitizedInput
-	}
-
-	/**
-	 * Add order items as a searchable string.
-	 *
-	 * This mimics how WooCommerce currently does in the order_itemmeta
-	 * table. They combine the titles of the products and put them in a
-	 * meta field called "Items".
-	 *
-	 * @since 2.4
-	 *
-	 * @param array      $post_args Post arguments
-	 * @param string|int $post_id Post id
-	 *
-	 * @return array
-	 */
-	public function add_order_items_search( $post_args, $post_id ) {
-		$searchable_post_types = $this->get_admin_searchable_post_types();
-
-		// Make sure it is only WooCommerce orders we touch.
-		if ( ! in_array( $post_args['post_type'], $searchable_post_types, true ) ) {
-			return $post_args;
-		}
-
-		$post_indexable = Indexables::factory()->get( 'post' );
-
-		// Get order items.
-		$order     = wc_get_order( $post_id );
-		$item_meta = [];
-		foreach ( $order->get_items() as $delta => $product_item ) {
-			// WooCommerce 3.x uses WC_Order_Item_Product instance while 2.x an array
-			if ( is_object( $product_item ) && method_exists( $product_item, 'get_name' ) ) {
-				$item_meta['_items'][] = $product_item->get_name( 'edit' );
-			} elseif ( is_array( $product_item ) && isset( $product_item['name'] ) ) {
-				$item_meta['_items'][] = $product_item['name'];
-			}
-		}
-
-		// Prepare order items.
-		$item_meta['_items'] = empty( $item_meta['_items'] ) ? '' : implode( '|', $item_meta['_items'] );
-		$post_args['meta']   = array_merge( $post_args['meta'], $post_indexable->prepare_meta_types( $item_meta ) );
-
-		return $post_args;
-	}
-
-	/**
-	 * Add WooCommerce Product Attributes to EP Facets.
-	 *
-	 * @param array $taxonomies Taxonomies array
-	 * @return array
-	 */
-	public function add_product_attributes( $taxonomies = [] ) {
-		$attribute_names = wc_get_attribute_taxonomy_names();
-
-		foreach ( $attribute_names as $name ) {
-			if ( ! taxonomy_exists( $name ) ) {
-				continue;
-			}
-			$taxonomies[ $name ] = get_taxonomy( $name );
-		}
-
-		return $taxonomies;
-	}
-
-	/**
-	 * Add WooCommerce Fields to the Weighting Dashboard.
-	 *
-	 * @since 3.x
-	 *
-	 * @param array  $fields    Current weighting fields.
-	 * @param string $post_type Current post type.
-	 * @return array            New fields.
-	 */
-	public function add_product_attributes_to_weighting( $fields, $post_type ) {
-		if ( 'product' === $post_type ) {
-			if ( ! empty( $fields['attributes']['children']['author_name'] ) ) {
-				unset( $fields['attributes']['children']['author_name'] );
-			}
-
-			$sku_key = 'meta._sku.value';
-
-			$fields['attributes']['children'][ $sku_key ] = array(
-				'key'   => $sku_key,
-				'label' => __( 'SKU', 'elasticpress' ),
-			);
-
-			$variations_skus_key = 'meta._variations_skus.value';
-
-			$fields['attributes']['children'][ $variations_skus_key ] = array(
-				'key'   => $variations_skus_key,
-				'label' => __( 'Variations SKUs', 'elasticpress' ),
-			);
-		}
-		return $fields;
-	}
-
-	/**
-	 * Add WooCommerce Fields to the default values of the Weighting Dashboard.
-	 *
-	 * @since 3.x
-	 *
-	 * @param array  $defaults  Default values for the post type.
-	 * @param string $post_type Current post type.
-	 * @return array
-	 */
-	public function add_product_default_post_type_weights( $defaults, $post_type ) {
-		if ( 'product' === $post_type ) {
-			if ( ! empty( $defaults['author_name'] ) ) {
-				unset( $defaults['author_name'] );
-			}
-
-			$defaults['meta._sku.value'] = array(
-				'enabled' => true,
-				'weight'  => 1,
-			);
-
-			$defaults['meta._variations_skus.value'] = array(
-				'enabled' => true,
-				'weight'  => 1,
-			);
-		}
-		return $defaults;
-	}
-
-	/**
-	 * Add WC post type to autosuggest
-	 *
-	 * @param array $post_types Array of post types (e.g. post, page).
-	 * @since  2.6
-	 * @return array
-	 */
-	public function suggest_wc_add_post_type( $post_types ) {
-		if ( ! in_array( 'product', $post_types, true ) ) {
-			$post_types[] = 'product';
-		}
-
-		return $post_types;
-	}
-
-	/**
-	 * Setup all feature filters
-	 *
-	 * @since  2.1
-	 */
-	public function setup() {
-		if ( ! function_exists( 'WC' ) ) {
-			return;
-		}
-
-		add_action( 'ep_formatted_args', [ $this, 'price_filter' ], 10, 3 );
-		add_filter( 'ep_sync_insert_permissions_bypass', [ $this, 'bypass_order_permissions_check' ], 10, 2 );
-		add_filter( 'ep_integrate_search_queries', [ $this, 'blacklist_coupons' ], 10, 2 );
-		add_filter( 'ep_prepare_meta_allowed_protected_keys', [ $this, 'whitelist_meta_keys' ], 10, 2 );
-		add_filter( 'woocommerce_layered_nav_query_post_ids', [ $this, 'convert_post_object_to_id' ], 10, 4 );
-		add_filter( 'woocommerce_unfiltered_product_ids', [ $this, 'convert_post_object_to_id' ], 10, 4 );
-		add_filter( 'ep_sync_taxonomies', [ $this, 'whitelist_taxonomies' ], 10, 2 );
-		add_filter( 'ep_post_sync_args_post_prepare_meta', [ $this, 'add_order_items_search' ], 20, 2 );
-		add_filter( 'ep_pc_skip_post_content_cleanup', [ $this, 'keep_order_fields' ], 20, 2 );
-		add_action( 'pre_get_posts', [ $this, 'translate_args' ], 11, 1 );
-		add_action( 'ep_wp_query_search_cached_posts', [ $this, 'disallow_duplicated_query' ], 10, 2 );
-		add_action( 'parse_query', [ $this, 'maybe_hook_woocommerce_search_fields' ], 1 );
-		add_action( 'parse_query', [ $this, 'search_order' ], 11 );
-		add_filter( 'ep_term_suggest_post_type', [ $this, 'suggest_wc_add_post_type' ] );
-		add_filter( 'ep_facet_include_taxonomies', [ $this, 'add_product_attributes' ] );
-		add_filter( 'ep_weighting_fields_for_post_type', [ $this, 'add_product_attributes_to_weighting' ], 10, 2 );
-		add_filter( 'ep_weighting_default_post_type_weights', [ $this, 'add_product_default_post_type_weights' ], 10, 2 );
-		add_filter( 'ep_prepare_meta_data', [ $this, 'add_variations_skus_meta' ], 10, 2 );
-		add_filter( 'request', [ $this, 'admin_product_list_request_query' ], 9 );
-
-		// Custom product ordering
-		add_action( 'ep_admin_notices', [ $this, 'maybe_display_notice_about_product_ordering' ] );
-		add_action( 'woocommerce_after_product_ordering', [ $this, 'action_sync_on_woocommerce_sort_single' ], 10, 2 );
-
-		// Orders Autosuggest feature.
-		if ( $this->is_orders_autosuggest_enabled() ) {
-			$this->orders->setup();
-		}
-
-		// Add WooCommerce Settings for Weight results by date
-		add_action( 'ep_weight_settings_after_search', [ $this, 'add_weight_settings_search' ] );
-		// Modify decaying based on WooCommerce Settings
-		add_filter( 'ep_is_decaying_enabled', [ $this, 'maybe_disable_decaying' ], 10, 3 );
 	}
 
 	/**
@@ -981,265 +578,13 @@ class WooCommerce extends Feature {
 	}
 
 	/**
-	 * Modifies main query to allow filtering by price with WooCommerce "Filter by price" widget.
-	 *
-	 * @param  array    $args ES args
-	 * @param  array    $query_args WP_Query args
-	 * @param  WP_Query $query WP_Query object
-	 * @since  3.2
-	 * @return array
-	 */
-	public function price_filter( $args, $query_args, $query ) {
-		// Only can use widget on main query
-		if ( ! $query->is_main_query() ) {
-			return $args;
-		}
-
-		// Only can use widget on shop, product taxonomy, or search
-		if ( ! is_shop() && ! is_product_taxonomy() && ! is_search() ) {
-			return $args;
-		}
-
-		// phpcs:disable WordPress.Security.NonceVerification
-		if ( empty( $_GET['min_price'] ) && empty( $_GET['max_price'] ) ) {
-			return $args;
-		}
-
-		$min_price = ! empty( $_GET['min_price'] ) ? sanitize_text_field( wp_unslash( $_GET['min_price'] ) ) : null;
-		$max_price = ! empty( $_GET['max_price'] ) ? sanitize_text_field( wp_unslash( $_GET['max_price'] ) ) : null;
-		// phpcs:enable WordPress.Security.NonceVerification
-
-		if ( $query->is_search() ) {
-			/**
-			 * This logic is iffy but the WC price filter widget is not intended for use with search anyway
-			 */
-			$old_query = $args['query']['bool'];
-			unset( $args['query']['bool']['should'] );
-
-			if ( ! empty( $min_price ) ) {
-				$args['query']['bool']['must'][0]['range']['meta._price.long']['gte'] = $min_price;
-			}
-
-			if ( ! empty( $max_price ) ) {
-				$args['query']['bool']['must'][0]['range']['meta._price.long']['lte'] = $max_price;
-			}
-
-			$args['query']['bool']['must'][0]['range']['meta._price.long']['boost'] = 2.0;
-			$args['query']['bool']['must'][1]['bool']                               = $old_query;
-		} else {
-			unset( $args['query']['match_all'] );
-
-			$args['query']['range']['meta._price.long']['gte'] = ! empty( $min_price ) ? $min_price : 0;
-
-			if ( ! empty( $min_price ) ) {
-				$args['query']['range']['meta._price.long']['gte'] = $min_price;
-			}
-
-			if ( ! empty( $max_price ) ) {
-				$args['query']['range']['meta._price.long']['lte'] = $max_price;
-			}
-
-			$args['query']['range']['meta._price.long']['boost'] = 2.0;
-		}
-
-		return $args;
-	}
-
-	/**
-	 * Prevent order fields from being removed.
-	 *
-	 * When Protected Content is enabled, all posts with password have their content removed.
-	 * This can't happen for orders, as the order key is added in that field.
-	 *
-	 * @see https://github.com/10up/ElasticPress/issues/2726
-	 *
-	 * @since 4.2.0
-	 * @param bool  $skip      Whether the password protected content should have their content, and meta removed
-	 * @param array $post_args Post arguments
-	 * @return bool
-	 */
-	public function keep_order_fields( $skip, $post_args ) {
-		$searchable_post_types = $this->get_admin_searchable_post_types();
-
-		if ( in_array( $post_args['post_type'], $searchable_post_types, true ) ) {
-			return true;
-		}
-
-		return $skip;
-	}
-
-	/**
-	 * Add a new `_variations_skus` meta field to the product to be indexed in Elasticsearch.
-	 *
-	 * @since 4.2.0
-	 * @param array   $post_meta Post meta
-	 * @param WP_Post $post      Post object
-	 * @return array
-	 */
-	public function add_variations_skus_meta( $post_meta, $post ) {
-		if ( 'product' !== $post->post_type ) {
-			return $post_meta;
-		}
-
-		$product = wc_get_product( $post );
-		if ( ! $product ) {
-			return $post_meta;
-		}
-
-		$variations_ids = $product->get_children();
-
-		$post_meta['_variations_skus'] = array_reduce(
-			$variations_ids,
-			function ( $variations_skus, $current_id ) {
-				$variation = wc_get_product( $current_id );
-				if ( ! $variation || ! $variation->exists() ) {
-					return $variations_skus;
-				}
-				$variation_sku = $variation->get_sku();
-				if ( ! $variation_sku ) {
-					return $variations_skus;
-				}
-				$variations_skus[] = $variation_sku;
-				return $variations_skus;
-			},
-			[]
-		);
-
-		return $post_meta;
-	}
-
-	/**
-	 * Integrate ElasticPress with the WooCommerce Admin Product List.
-	 *
-	 * WooCommerce uses its `WC_Admin_List_Table_Products` class to control that screen. This
-	 * function adds all necessary hooks to bypass the default behavior and integrate with ElasticPress.
-	 * By default, WC runs a SQL query to get the Product IDs that match the list criteria and passes
-	 * that list of IDs to the main WP_Query. This integration changes that process to a single query, run
-	 * by ElasticPress.
-	 *
-	 * @since 4.2.0
-	 * @param array $query_vars Query vars.
-	 * @return array
-	 */
-	public function admin_product_list_request_query( $query_vars ) {
-		global $typenow, $wc_list_table;
-
-		// Return if not in the correct screen.
-		if ( ! is_a( $wc_list_table, 'WC_Admin_List_Table_Products' ) || 'product' !== $typenow ) {
-			return $query_vars;
-		}
-
-		// Return if admin WP_Query integration is not turned on, i.e., Protect Content is not enabled.
-		if ( ! has_filter( 'ep_admin_wp_query_integration', '__return_true' ) ) {
-			return $query_vars;
-		}
-
-		/**
-		 * Filter to skip integration with WooCommerce Admin Product List.
-		 *
-		 * @hook ep_woocommerce_integrate_admin_products_list
-		 * @since 4.2.0
-		 * @param {bool}  $integrate  True to integrate, false to preserve original behavior. Defaults to true.
-		 * @param {array} $query_vars Query vars.
-		 * @return {bool} New integrate value
-		 */
-		if ( ! apply_filters( 'ep_woocommerce_integrate_admin_products_list', true, $query_vars ) ) {
-			return $query_vars;
-		}
-
-		add_action( 'pre_get_posts', [ $this, 'translate_args_admin_products_list' ], 12 );
-
-		// This short-circuits WooCommerce search for product IDs.
-		add_filter( 'woocommerce_product_pre_search_products', '__return_empty_array' );
-
-		return $query_vars;
-	}
-
-	/**
-	 * Apply the necessary changes to WP_Query in WooCommerce Admin Product List.
-	 *
-	 * @param WP_Query $query The WP Query being executed.
-	 */
-	public function translate_args_admin_products_list( $query ) {
-		// The `translate_args()` method sets it to `true` if we should integrate it.
-		if ( ! $query->get( 'ep_integrate', false ) ) {
-			return;
-		}
-
-		// WooCommerce unsets the search term right after using it to fetch product IDs. Here we add it back.
-		$search_term = ! empty( $_GET['s'] ) ? sanitize_text_field( wp_unslash( $_GET['s'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification
-		if ( ! empty( $search_term ) ) {
-			$query->set( 's', sanitize_text_field( $search_term ) ); // phpcs:ignore WordPress.Security.NonceVerification
-
-			/**
-			 * Filter the fields used in WooCommerce Admin Product Search.
-			 *
-			 * ```
-			 * add_filter(
-			 *     'ep_woocommerce_admin_products_list_search_fields',
-			 *     function ( $wc_admin_search_fields ) {
-			 *         $wc_admin_search_fields['meta'][] = 'custom_field';
-			 *         return $wc_admin_search_fields;
-			 *     }
-			 * );
-			 * ```
-			 *
-			 * @hook ep_woocommerce_admin_products_list_search_fields
-			 * @since 4.2.0
-			 * @param {array} $wc_admin_search_fields Fields to be used in the WooCommerce Admin Product Search
-			 * @return {array} New fields
-			 */
-			$search_fields = apply_filters(
-				'ep_woocommerce_admin_products_list_search_fields',
-				[
-					'post_title',
-					'post_content',
-					'post_excerpt',
-					'meta' => [
-						'_sku',
-						'_variations_skus',
-					],
-				]
-			);
-
-			$query->set( 'search_fields', $search_fields );
-		}
-
-		// Sets the meta query for `product_type` if needed. Also removed from the WP_Query by WC in `WC_Admin_List_Table_Products::query_filters()`.
-		$product_type_query = $query->get( 'product_type', '' );
-		$product_type_url   = ! empty( $_GET['product_type'] ) ? sanitize_text_field( wp_unslash( $_GET['product_type'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification
-		$allowed_prod_types = [ 'virtual', 'downloadable' ];
-		if ( empty( $product_type_query ) && ! empty( $product_type_url ) && in_array( $product_type_url, $allowed_prod_types, true ) ) {
-			$meta_query   = $query->get( 'meta_query', [] );
-			$meta_query[] = [
-				'key'   => "_{$product_type_url}",
-				'value' => 'yes',
-			];
-			$query->set( 'meta_query', $meta_query );
-		}
-
-		// Sets the meta query for `stock_status` if needed.
-		$stock_status_query   = $query->get( 'stock_status', '' );
-		$stock_status_url     = ! empty( $_GET['stock_status'] ) ? sanitize_text_field( wp_unslash( $_GET['stock_status'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification
-		$allowed_stock_status = [ 'instock', 'outofstock', 'onbackorder' ];
-		if ( empty( $stock_status_query ) && ! empty( $stock_status_url ) && in_array( $stock_status_url, $allowed_stock_status, true ) ) {
-			$meta_query   = $query->get( 'meta_query', [] );
-			$meta_query[] = [
-				'key'   => '_stock_status',
-				'value' => $stock_status_url,
-			];
-			$query->set( 'meta_query', $meta_query );
-		}
-	}
-
-	/**
 	 * Determines whether or not ES should be integrating with the provided query
 	 *
 	 * @param \WP_Query $query Query we might integrate with
 	 *
 	 * @return bool
 	 */
-	protected function should_integrate_with_query( $query ) {
+	public function should_integrate_with_query( $query ) {
 		// Lets make sure this doesn't interfere with the CLI
 		if ( defined( 'WP_CLI' ) && WP_CLI ) {
 			return false;
@@ -1296,61 +641,6 @@ class WooCommerce extends Feature {
 	}
 
 	/**
-	 * Depending on the number of products display an admin notice in the custom sort screen for WooCommerce Products
-	 *
-	 * @since 4.4.0
-	 * @param array $notices Current ElasticPress admin notices
-	 * @return array
-	 */
-	public function maybe_display_notice_about_product_ordering( $notices ) {
-		global $pagenow, $wp_query;
-
-		/**
-		 * Make sure we're on edit.php in admin dashboard.
-		 */
-		if ( ! is_admin() || 'edit.php' !== $pagenow || empty( $wp_query->query['orderby'] ) || 'menu_order title' !== $wp_query->query['orderby'] ) {
-			return $notices;
-		}
-
-		$documents_per_page_sync = IndexHelper::factory()->get_index_default_per_page();
-		if ( $documents_per_page_sync >= $wp_query->found_posts ) {
-			return $notices;
-		}
-
-		$notices['woocommerce_custom_sort'] = [
-			'html'    => sprintf(
-				/* translators: Sync Page URL */
-				__( 'Due to the number of products in the site, you will need to <a href="%s">resync</a> after applying a custom sort order.', 'elasticpress' ),
-				Utils\get_sync_url()
-			),
-			'type'    => 'warning',
-			'dismiss' => true,
-		];
-
-		return $notices;
-	}
-
-	/**
-	 * Conditionally resync products after applying a custom order.
-	 *
-	 * @since 4.4.0
-	 * @param int   $sorting_id  ID of post dragged and dropped
-	 * @param array $menu_orders Post IDs and their new menu_order value
-	 */
-	public function action_sync_on_woocommerce_sort_single( $sorting_id, $menu_orders ) {
-
-		$documents_per_page_sync = IndexHelper::factory()->get_index_default_per_page();
-		if ( $documents_per_page_sync < count( $menu_orders ) ) {
-			return;
-		}
-
-		$sync_manager = Indexables::factory()->get( 'post' )->sync_manager;
-		foreach ( $menu_orders as $post_id => $order ) {
-			$sync_manager->add_to_queue( $post_id );
-		}
-	}
-
-	/**
 	 * Whether orders autosuggest is available or not
 	 *
 	 * @since 4.5.0
@@ -1378,21 +668,314 @@ class WooCommerce extends Feature {
 		return $this->is_orders_autosuggest_available() && '1' === $this->get_setting( 'orders' );
 	}
 
+
 	/**
-	 * Add weight by date settings related to WooCommerce
+	 * DEPRECATED. Index Woocommerce meta
+	 *
+	 * @param   array $meta Existing post meta.
+	 * @param   array $post Post arguments array.
+	 * @since   2.1
+	 * @return  array
+	 */
+	public function whitelist_meta_keys( $meta, $post ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->allow_meta_keys() AND/OR \ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->allow_meta_keys()" );
+		return array_unique(
+			array_merge(
+				$this->products->allow_meta_keys( $meta ),
+				$this->orders->allow_meta_keys( $meta )
+			)
+		);
+	}
+
+	/**
+	 * DEPRECATED. Make sure all loop shop post ins are IDS. We have to pass post objects here since we override
+	 * the fields=>id query for the layered filter nav query
+	 *
+	 * @param   array $posts Post object array.
+	 * @since   2.1
+	 * @return  array
+	 */
+	public function convert_post_object_to_id( $posts ) {
+		_doing_it_wrong( __METHOD__, 'This filter was removed from WooCommerce and will be removed from ElasticPress in a future release.', '4.5.0' );
+		return $posts;
+	}
+
+	/**
+	 * DEPRECATED. Index Woocommerce taxonomies
+	 *
+	 * @param   array $taxonomies Index taxonomies array.
+	 * @param   array $post Post properties array.
+	 * @since   2.1
+	 * @return  array
+	 */
+	public function whitelist_taxonomies( $taxonomies, $post ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->sync_taxonomies()" );
+		return $this->products->sync_taxonomies( $taxonomies );
+	}
+
+	/**
+	 * DEPRECATED. Disallow duplicated ES queries on Orders page.
+	 *
+	 * @since 2.4
+	 *
+	 * @param array    $value Original filter values.
+	 * @param WP_Query $query WP_Query
+	 *
+	 * @return array
+	 */
+	public function disallow_duplicated_query( $value, $query ) {
+		_doing_it_wrong( __METHOD__, 'This filter was removed from WooCommerce and will be removed from ElasticPress in a future release.', '4.5.0' );
+
+		return $value;
+	}
+
+	/**
+	 * DEPRECATED. Returns the WooCommerce-oriented post types in admin that EP will search
+	 *
+	 * @since 4.4.0
+	 * @return mixed|void
+	 */
+	public function get_admin_searchable_post_types() {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->get_admin_searchable_post_types()" );
+		return $this->orders->get_admin_searchable_post_types();
+	}
+
+	/**
+	 * DEPRECATED. Make search coupons don't go through ES
+	 *
+	 * @param  bool     $enabled Coupons enabled or not
+	 * @param  WP_Query $query WP Query
+	 * @since  2.1
+	 * @return bool
+	 */
+	public function blacklist_coupons( $enabled, $query ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->disallow_coupons()" );
+		return $this->disallow_coupons( $enabled, $query );
+	}
+
+	/**
+	 * DEPRECATED. Allow order creations on the front end to get synced
+	 *
+	 * @since  2.1
+	 * @param  bool $override Original order perms check value
+	 * @param  int  $post_id Post ID
+	 * @return bool
+	 */
+	public function bypass_order_permissions_check( $override, $post_id ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->price_filter()" );
+		return $this->orders->bypass_order_permissions_check( $override, $post_id );
+	}
+
+	/**
+	 * DEPRECATED. Sets woocommerce meta search fields to an empty array if we are integrating the main query with ElasticSearch
+	 *
+	 * Woocommerce calls this action as part of its own callback on parse_query. We add this filter only if the query
+	 * is integrated with ElasticSearch.
+	 * If we were to always return array() on this filter, we'd break admin searches when WooCommerce module is activated
+	 * without the Protected Content Module
+	 *
+	 * @param \WP_Query $query Current query
+	 */
+	public function maybe_hook_woocommerce_search_fields( $query ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->maybe_hook_woocommerce_search_fields()" );
+		return $this->orders->maybe_hook_woocommerce_search_fields( $query );
+	}
+
+	/**
+	 * DEPRECATED. Enhance WooCommerce search order by order id, email, phone number, name, etc..
+	 * What this function does:
+	 * 1. Reverse the woocommerce shop_order_search_custom_fields query
+	 * 2. If the search key is integer and it is an Order Id, just query with post__in
+	 * 3. If the search key is integer but not an order id ( might be phone number ), use ES to find it
+	 *
+	 * @param WP_Query $wp WP Query
+	 * @since  2.3
+	 */
+	public function search_order( $wp ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->search_order()" );
+		return $this->orders->search_order( $wp );
+	}
+
+	/**
+	 * DEPRECATED. Add order items as a searchable string.
+	 *
+	 * This mimics how WooCommerce currently does in the order_itemmeta
+	 * table. They combine the titles of the products and put them in a
+	 * meta field called "Items".
+	 *
+	 * @since 2.4
+	 *
+	 * @param array      $post_args Post arguments
+	 * @param string|int $post_id Post id
+	 *
+	 * @return array
+	 */
+	public function add_order_items_search( $post_args, $post_id ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->add_order_items_search()" );
+		return $this->orders->add_order_items_search( $post_args, $post_id );
+	}
+
+	/**
+	 * DEPRECATED. Add WooCommerce Product Attributes to EP Facets.
+	 *
+	 * @param array $taxonomies Taxonomies array
+	 * @return array
+	 */
+	public function add_product_attributes( $taxonomies = [] ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->add_product_attributes()" );
+		return $this->products->add_product_attributes( $taxonomies );
+	}
+
+	/**
+	 * DEPRECATED. Add WooCommerce Fields to the Weighting Dashboard.
+	 *
+	 * @since 3.x
+	 *
+	 * @param array  $fields    Current weighting fields.
+	 * @param string $post_type Current post type.
+	 * @return array            New fields.
+	 */
+	public function add_product_attributes_to_weighting( $fields, $post_type ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->add_product_attributes_to_weighting()" );
+		return $this->products->add_product_attributes_to_weighting( $fields, $post_type );
+	}
+
+	/**
+	 * DEPRECATED. Add WooCommerce Fields to the default values of the Weighting Dashboard.
+	 *
+	 * @since 3.x
+	 *
+	 * @param array  $defaults  Default values for the post type.
+	 * @param string $post_type Current post type.
+	 * @return array
+	 */
+	public function add_product_default_post_type_weights( $defaults, $post_type ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->add_product_default_post_type_weights()" );
+		return $this->products->add_product_default_post_type_weights( $defaults, $post_type );
+	}
+
+	/**
+	 * DEPRECATED. Add WC post type to autosuggest
+	 *
+	 * @param array $post_types Array of post types (e.g. post, page).
+	 * @since  2.6
+	 * @return array
+	 */
+	public function suggest_wc_add_post_type( $post_types ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->suggest_wc_add_post_type()" );
+		return $this->products->suggest_wc_add_post_type( $post_types );
+	}
+
+	/**
+	 * DEPRECATED. Modifies main query to allow filtering by price with WooCommerce "Filter by price" widget.
+	 *
+	 * @param  array    $args ES args
+	 * @param  array    $query_args WP_Query args
+	 * @param  WP_Query $query WP_Query object
+	 * @since  3.2
+	 * @return array
+	 */
+	public function price_filter( $args, $query_args, $query ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->price_filter()" );
+		return $this->products->price_filter( $args, $query_args, $query );
+	}
+
+	/**
+	 * DEPRECATED. Prevent order fields from being removed.
+	 *
+	 * When Protected Content is enabled, all posts with password have their content removed.
+	 * This can't happen for orders, as the order key is added in that field.
+	 *
+	 * @see https://github.com/10up/ElasticPress/issues/2726
+	 *
+	 * @since 4.2.0
+	 * @param bool  $skip      Whether the password protected content should have their content, and meta removed
+	 * @param array $post_args Post arguments
+	 * @return bool
+	 */
+	public function keep_order_fields( $skip, $post_args ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->orders->keep_order_fields()" );
+		return $this->orders->keep_order_fields( $skip, $post_args );
+	}
+
+	/**
+	 * DEPRECATED. Add a new `_variations_skus` meta field to the product to be indexed in Elasticsearch.
+	 *
+	 * @since 4.2.0
+	 * @param array   $post_meta Post meta
+	 * @param WP_Post $post      Post object
+	 * @return array
+	 */
+	public function add_variations_skus_meta( $post_meta, $post ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->add_variations_skus_meta()" );
+		return $this->products->add_variations_skus_meta( $post_meta, $post );
+	}
+
+	/**
+	 * DEPRECATED. Integrate ElasticPress with the WooCommerce Admin Product List.
+	 *
+	 * WooCommerce uses its `WC_Admin_List_Table_Products` class to control that screen. This
+	 * function adds all necessary hooks to bypass the default behavior and integrate with ElasticPress.
+	 * By default, WC runs a SQL query to get the Product IDs that match the list criteria and passes
+	 * that list of IDs to the main WP_Query. This integration changes that process to a single query, run
+	 * by ElasticPress.
+	 *
+	 * @since 4.2.0
+	 * @param array $query_vars Query vars.
+	 * @return array
+	 */
+	public function admin_product_list_request_query( $query_vars ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->admin_product_list_request_query()" );
+		return $this->products->admin_product_list_request_query( $query_vars );
+	}
+
+	/**
+	 * DEPRECATED. Apply the necessary changes to WP_Query in WooCommerce Admin Product List.
+	 *
+	 * @param WP_Query $query The WP Query being executed.
+	 */
+	public function translate_args_admin_products_list( $query ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->price_filter()" );
+		$this->products->translate_args_admin_products_list( $query );
+	}
+
+	/**
+	 * DEPRECATED. Depending on the number of products display an admin notice in the custom sort screen for WooCommerce Products
+	 *
+	 * @since 4.4.0
+	 * @param array $notices Current ElasticPress admin notices
+	 * @return array
+	 */
+	public function maybe_display_notice_about_product_ordering( $notices ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->maybe_display_notice_about_product_ordering()" );
+		return $this->products->maybe_display_notice_about_product_ordering( $notices );
+	}
+
+	/**
+	 * DEPRECATED. Conditionally resync products after applying a custom order.
+	 *
+	 * @since 4.4.0
+	 * @param int   $sorting_id  ID of post dragged and dropped
+	 * @param array $menu_orders Post IDs and their new menu_order value
+	 */
+	public function action_sync_on_woocommerce_sort_single( $sorting_id, $menu_orders ) {
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->action_sync_on_woocommerce_sort_single()" );
+		return $this->products->action_sync_on_woocommerce_sort_single( $sorting_id, $menu_orders );
+	}
+
+	/**
+	 * DEPRECATED. Add weight by date settings related to WooCommerce
 	 *
 	 * @since 4.6.0
 	 * @param array $settings Current settings.
 	 */
 	public function add_weight_settings_search( $settings ) {
-		?>
-		<label><input name="settings[decaying_enabled]" type="radio" <?php checked( $settings['decaying_enabled'], 'disabled_only_products' ); ?> value="disabled_only_products"><?php esc_html_e( 'Disabled for product only queries', 'elasticpress' ); ?></label><br>
-		<label><input name="settings[decaying_enabled]" type="radio" <?php checked( $settings['decaying_enabled'], 'disabled_includes_products' ); ?> value="disabled_includes_products"><?php esc_html_e( 'Disabled for any query that includes products', 'elasticpress' ); ?></label>
-		<?php
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->add_weight_settings_search()" );
+		$this->products->add_weight_settings_search( $settings );
 	}
 
 	/**
-	 * Conditionally disable decaying by date based on WooCommerce Decay settings.
+	 * DEPRECATED. Conditionally disable decaying by date based on WooCommerce Decay settings.
 	 *
 	 * @since 4.6.0
 	 * @param bool  $is_decaying_enabled Whether decay by date is enabled or not
@@ -1401,25 +984,7 @@ class WooCommerce extends Feature {
 	 * @return bool
 	 */
 	public function maybe_disable_decaying( $is_decaying_enabled, $settings, $args ) {
-		if ( ! in_array( $settings['decaying_enabled'], [ 'disabled_only_products', 'disabled_includes_products' ], true ) ) {
-			return $is_decaying_enabled;
-		}
-
-		if ( ! isset( $args['post_type'] ) || ! in_array( 'product', (array) $args['post_type'], true ) ) {
-			return $is_decaying_enabled;
-		}
-
-		$post_types = (array) $args['post_type'];
-
-		if ( 'disabled_only_products' === $settings['decaying_enabled'] && count( $post_types ) > 1 ) {
-			return $is_decaying_enabled;
-		}
-
-		if ( 'disabled_includes_products' === $settings['decaying_enabled'] && ! in_array( 'product', $post_types, true ) ) {
-			return $is_decaying_enabled;
-		}
-
-		return false;
+		_deprecated_function( __METHOD__, '4.7.0', "\ElasticPress\Features::factory()->get_registered_feature( 'woocommerce' )->products->maybe_disable_decaying()" );
+		return $this->products->maybe_disable_decaying( $is_decaying_enabled, $settings, $args );
 	}
-
 }
